Fixed paths for imports after moving tests
